+++ src/Castle.Core.Tests/DynamicProxy.Tests/ClassWithAttributesTestCase.cs
@@ -113,6 +113,53 @@ namespace Castle.DynamicProxy.Tests
 		}
 
 		[Test]
+		public void EnsureProxyHasNestedAttributesOnClassAndMethods()
+		{
+			var instance = (HasNestedNonInheritableAttribute)generator.CreateClassProxy(typeof(HasNestedNonInheritableAttribute), new StandardInterceptor());
+
+			object[] attributes = instance.GetType().GetCustomAttributes(typeof(Nested.NonInheritableAttribute), false).ToArray();
+			Assert.AreEqual(1, attributes.Length);
+			Assert.IsInstanceOf(typeof(Nested.NonInheritableAttribute), attributes[0]);
+
+			attributes = instance.GetType().GetMethod("OnMethod").GetCustomAttributes(typeof(Nested.NonInheritableAttribute), false).ToArray();
+			Assert.AreEqual(1, attributes.Length);
+			Assert.IsInstanceOf(typeof(Nested.NonInheritableAttribute), attributes[0]);
+		}
+
+		[Test]
+		public void EnsureProxyHasNestedAttributesOnProperties()
+		{
+			var proxy = generator.CreateClassProxy<HasNestedNonInheritableAttribute>();
+			var nameProperty = proxy.GetType().GetProperty("OnProperty");
+			Assert.IsTrue(nameProperty.IsDefined(typeof(Nested.NonInheritableAttribute), false));
+		}
+
+		[Test, Ignore("Not supported. Is it possible? There seems to be no API to allow that.")]
+		public void EnsureProxyHasNestedAttributesOnOnReturn()
+		{
+			var proxy = generator.CreateClassProxy<HasNestedNonInheritableAttribute>();
+			var nameProperty = proxy.GetType().GetMethod("OnReturn").ReturnParameter;
+			Assert.IsTrue(nameProperty.IsDefined(typeof(Nested.NonInheritableAttribute), false));
+		}
+
+		[Test]
+		public void EnsureProxyHasNestedAttributesOnParameter()
+		{
+			var proxy = generator.CreateClassProxy<HasNestedNonInheritableAttribute>();
+			ParameterInfo nameProperty = proxy.GetType().GetMethod("OnParameter").GetParameters().Single();
+			Assert.IsTrue(nameProperty.IsDefined(typeof(Nested.NonInheritableAttribute), false));
+		}
+
+		[Test]
+		[Platform(Exclude = "Mono", Reason = "Mono does not currently emit custom attributes on generic type parameters of methods. See https://github.com/mono/mono/issues/8512.")]
+		public void EnsureProxyHasNestedAttributesOnGenericArgument()
+		{
+			var proxy = generator.CreateClassProxy<HasNestedNonInheritableAttribute>();
+			var nameProperty = proxy.GetType().GetMethod("OnGenericArgument").GetGenericArguments().Single();
+			Assert.IsTrue(nameProperty.IsDefined(typeof(Nested.NonInheritableAttribute), false));
+		}
+
+		[Test]
 		public void Can_proxy_type_with_non_inheritable_attribute_depending_on_array_of_something_via_property()
 		{
 			var proxy = generator.CreateInterfaceProxyWithoutTarget<IHasNonInheritableAttributeWithArray>();